generic gmp3 build for m68k;compiler changes to declare all functions as returning object, with functions that actually return long being cast appropriately;back out of m68k hack in eval.c and funlink.c

+++ gcl/lsp/destructuring_bind.c
@@ -494,7 +494,7 @@ T213:;
 goto T219;
 T219:;
 	if((V46!= VV[27]))goto T240;
-	if(!((length((V41)))>(3))){
+	if(!(((long)length((V41)))>(3))){
 	goto T241;}
 	base[8]= VV[35];
 	base[9]= VV[36];
